refactor: reorganize folders structure for services and pipes
