+++ src/app/simulados/components/exam-card/exam-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Simulado } from '../../utils/simulado';
+import { Simulado } from '../../utils/model/simulado';
 
 @Component({
   selector: 'app-exam-card',
