+++ src/app/simulados/pages/exam/exam.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import { Simulado } from '../../utils/simulado';
+import { Simulado } from '../../utils/model/simulado';
 import { ActivatedRoute, Router } from '@angular/router';
-import { QueryExamsService } from '../../utils/query-exams.service';
+import { QueryExamsService } from '../../utils/service/query-exams.service';
 
 @Component({
   selector: 'app-exam',
