+++ src/app/simulados/utils/directive/dinamic-loader.directive.spec.ts
