+++ src/app/simulados/utils/model/domain.ts
