+++ src/app/simulados/utils/model/question-interface.ts
