+++ src/app/simulados/utils/pipe/summary.pipe.ts
