+++ src/app/simulados/utils/service/query-code.service.spec.ts
