+++ src/app/simulados/utils/service/query-exams.service.ts
@@ -1,8 +1,8 @@
 import { lastValueFrom } from 'rxjs';
-import { Simulado } from './simulado';
+import { Simulado } from '../model/simulado';
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { environment } from '../../../environments/environment';
+import { environment } from '../../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
