+++ src/app/simulados/utils/service/simulado-events.service.ts
